requirements.txt
semgrep>=1.52.0
PyYAML>=6.0
